+++ R/eval.R
@@ -100,6 +100,9 @@
 #'   to returning the complete call stack, it also start up a debugger
 #'   in the child call stack, via [utils::debugger()].
 #'
+#' The default error behavior can be also set using the `callr.error`
+#' option. This is useful to debug code that uses `callr`.
+#'
 #' @family callr functions
 #' @examples
 #' \dontrun{
@@ -117,7 +120,7 @@ r <- function(func, args = list(), libpath = .libPaths(),
               repos = c(getOption("repos"),
                 c(CRAN = "https://cloud.r-project.org")),
               stdout = NULL, stderr = NULL,
-              error = c("error", "stack", "debugger"),
+              error = getOption("callr.error", "error"),
               cmdargs = c("--no-site-file", "--no-environ", "--slave",
                 "--no-save", "--no-restore"),
               show = FALSE, callback = NULL,